+++ spec/ObservableClass.spec.ts
@@ -85,48 +85,183 @@ describe("Observable Depth", () => {
     it("Deep", () => {
         let o = new Depth();
         o.deep = [{ a: "A" }];
-        expect(invalidates(() => o.deep, () => (o.deep = [{ a: "B1" }]))).toBeTruthy();
-        expect(invalidates(() => o.deep![0], () => (o.deep = [{ a: "B2" }]))).toBeTruthy();
-        expect(invalidates(() => o.deep![0].a, () => (o.deep = [{ a: "B3" }]))).toBeTruthy();
+        expect(
+            invalidates(
+                () => o.deep,
+                () => (o.deep = [{ a: "B1" }])
+            )
+        ).toBeTruthy();
+        expect(
+            invalidates(
+                () => o.deep![0],
+                () => (o.deep = [{ a: "B2" }])
+            )
+        ).toBeTruthy();
+        expect(
+            invalidates(
+                () => o.deep![0]!.a,
+                () => (o.deep = [{ a: "B3" }])
+            )
+        ).toBeTruthy();
 
-        expect(invalidates(() => o.deep, () => (o.deep![0] = { a: "B4" }))).toBeFalsy();
-        expect(invalidates(() => o.deep![0], () => (o.deep![0] = { a: "B5" }))).toBeTruthy();
-        expect(invalidates(() => o.deep![0].a, () => (o.deep![0] = { a: "B6" }))).toBeTruthy();
+        expect(
+            invalidates(
+                () => o.deep,
+                () => (o.deep![0] = { a: "B4" })
+            )
+        ).toBeFalsy();
+        expect(
+            invalidates(
+                () => o.deep![0],
+                () => (o.deep![0] = { a: "B5" })
+            )
+        ).toBeTruthy();
+        expect(
+            invalidates(
+                () => o.deep![0]!.a,
+                () => (o.deep![0] = { a: "B6" })
+            )
+        ).toBeTruthy();
 
-        expect(invalidates(() => o.deep, () => (o.deep![0].a = "B7"))).toBeFalsy();
-        expect(invalidates(() => o.deep![0], () => (o.deep![0].a = "B8"))).toBeFalsy();
-        expect(invalidates(() => o.deep![0].a, () => (o.deep![0].a = "B9"))).toBeTruthy();
+        expect(
+            invalidates(
+                () => o.deep,
+                () => (o.deep![0]!.a = "B7")
+            )
+        ).toBeFalsy();
+        expect(
+            invalidates(
+                () => o.deep![0],
+                () => (o.deep![0]!.a = "B8")
+            )
+        ).toBeFalsy();
+        expect(
+            invalidates(
+                () => o.deep![0]!.a,
+                () => (o.deep![0]!.a = "B9")
+            )
+        ).toBeTruthy();
     });
 
     it("Shallow", () => {
         let o = new Depth();
         o.shallow = [{ a: "A" }];
-        expect(invalidates(() => o.shallow, () => (o.shallow = [{ a: "B1" }]))).toBeTruthy();
-        expect(invalidates(() => o.shallow![0], () => (o.shallow = [{ a: "B2" }]))).toBeTruthy();
-        expect(invalidates(() => o.shallow![0].a, () => (o.shallow = [{ a: "B3" }]))).toBeTruthy();
+        expect(
+            invalidates(
+                () => o.shallow,
+                () => (o.shallow = [{ a: "B1" }])
+            )
+        ).toBeTruthy();
+        expect(
+            invalidates(
+                () => o.shallow![0],
+                () => (o.shallow = [{ a: "B2" }])
+            )
+        ).toBeTruthy();
+        expect(
+            invalidates(
+                () => o.shallow![0]!.a,
+                () => (o.shallow = [{ a: "B3" }])
+            )
+        ).toBeTruthy();
 
-        expect(invalidates(() => o.shallow, () => (o.shallow![0] = { a: "B4" }))).toBeFalsy();
-        expect(invalidates(() => o.shallow![0], () => (o.shallow![0] = { a: "B5" }))).toBeTruthy();
-        expect(invalidates(() => o.shallow![0].a, () => (o.shallow![0] = { a: "B6" }))).toBeTruthy();
+        expect(
+            invalidates(
+                () => o.shallow,
+                () => (o.shallow![0] = { a: "B4" })
+            )
+        ).toBeFalsy();
+        expect(
+            invalidates(
+                () => o.shallow![0],
+                () => (o.shallow![0] = { a: "B5" })
+            )
+        ).toBeTruthy();
+        expect(
+            invalidates(
+                () => o.shallow![0]!.a,
+                () => (o.shallow![0] = { a: "B6" })
+            )
+        ).toBeTruthy();
 
-        expect(invalidates(() => o.shallow, () => (o.shallow![0].a = "B7"))).toBeFalsy();
-        expect(invalidates(() => o.shallow![0], () => (o.shallow![0].a = "B8"))).toBeFalsy();
-        expect(invalidates(() => o.shallow![0].a, () => (o.shallow![0].a = "B9"))).toBeFalsy();
+        expect(
+            invalidates(
+                () => o.shallow,
+                () => (o.shallow![0]!.a = "B7")
+            )
+        ).toBeFalsy();
+        expect(
+            invalidates(
+                () => o.shallow![0],
+                () => (o.shallow![0]!.a = "B8")
+            )
+        ).toBeFalsy();
+        expect(
+            invalidates(
+                () => o.shallow![0]!.a,
+                () => (o.shallow![0]!.a = "B9")
+            )
+        ).toBeFalsy();
     });
 
     it("Ref", () => {
         let o = new Depth();
         o.ref = [{ a: "A" }];
-        expect(invalidates(() => o.ref, () => (o.ref = [{ a: "B1" }]))).toBeTruthy();
-        expect(invalidates(() => o.ref![0], () => (o.ref = [{ a: "B2" }]))).toBeTruthy();
-        expect(invalidates(() => o.ref![0].a, () => (o.ref = [{ a: "B3" }]))).toBeTruthy();
+        expect(
+            invalidates(
+                () => o.ref,
+                () => (o.ref = [{ a: "B1" }])
+            )
+        ).toBeTruthy();
+        expect(
+            invalidates(
+                () => o.ref![0],
+                () => (o.ref = [{ a: "B2" }])
+            )
+        ).toBeTruthy();
+        expect(
+            invalidates(
+                () => o.ref![0]!.a,
+                () => (o.ref = [{ a: "B3" }])
+            )
+        ).toBeTruthy();
 
-        expect(invalidates(() => o.ref, () => (o.ref![0] = { a: "B4" }))).toBeFalsy();
-        expect(invalidates(() => o.ref![0], () => (o.ref![0] = { a: "B5" }))).toBeFalsy();
-        expect(invalidates(() => o.ref![0].a, () => (o.ref![0] = { a: "B6" }))).toBeFalsy();
+        expect(
+            invalidates(
+                () => o.ref,
+                () => (o.ref![0] = { a: "B4" })
+            )
+        ).toBeFalsy();
+        expect(
+            invalidates(
+                () => o.ref![0],
+                () => (o.ref![0] = { a: "B5" })
+            )
+        ).toBeFalsy();
+        expect(
+            invalidates(
+                () => o.ref![0]!.a,
+                () => (o.ref![0] = { a: "B6" })
+            )
+        ).toBeFalsy();
 
-        expect(invalidates(() => o.ref, () => (o.ref![0].a = "B7"))).toBeFalsy();
-        expect(invalidates(() => o.ref![0], () => (o.ref![0].a = "B8"))).toBeFalsy();
-        expect(invalidates(() => o.ref![0].a, () => (o.ref![0].a = "B9"))).toBeFalsy();
+        expect(
+            invalidates(
+                () => o.ref,
+                () => (o.ref![0]!.a = "B7")
+            )
+        ).toBeFalsy();
+        expect(
+            invalidates(
+                () => o.ref![0],
+                () => (o.ref![0]!.a = "B8")
+            )
+        ).toBeFalsy();
+        expect(
+            invalidates(
+                () => o.ref![0]!.a,
+                () => (o.ref![0]!.a = "B9")
+            )
+        ).toBeFalsy();
     });
 });